Clamp invalid unix timestamp values to range allowed by DateTimeOffset

+++ WowPacketParser/Saving/FusionBinaryPacketWriter.cs
@@ -37,7 +37,7 @@ namespace WowPacketParser.Saving
                             writer.Write((ushort)packet.Opcode);
                             writer.Write((int)packet.Length);
                             writer.Write((byte)packet.Direction);
-                            writer.Write((ulong)Utilities.GetUnixTimeFromDateTime(packet.Time));
+                            writer.Write(Utilities.GetUnixTimeFromDateTime(packet.Time));
                             writer.Write(packet.GetStream(0));
                             // TODO: Add ConnIdx in a backwards compatible way
                         }
